x/lottery/keeper/grpc_query_round.go
package keeper

import (
	"context"

	"game/x/lottery/types"
	sdk "github.com/cosmos/cosmos-sdk/types"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"
)

func (k Keeper) Round(c context.Context, req *types.QueryGetRoundRequest) (*types.QueryGetRoundResponse, error) {
	if req == nil {
		return nil, status.Error(codes.InvalidArgument, "invalid request")
	}
	ctx := sdk.UnwrapSDKContext(c)

	val, found := k.GetRound(ctx)
	if !found {
		return nil, status.Error(codes.NotFound, "not found")
	}

	return &types.QueryGetRoundResponse{Round: val}, nil
}
